+++ src/OpenTelemetry/Metrics/MetricAggregators/GaugeMetricAggregator.cs
@@ -16,6 +16,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.Metrics;
 
 namespace OpenTelemetry.Metrics
 {
@@ -24,11 +25,17 @@ namespace OpenTelemetry.Metrics
         private readonly object lockUpdate = new object();
         private IDataValue value;
 
-        public GaugeMetricAggregator()
+        internal GaugeMetricAggregator()
         {
         }
 
         public string Name { get; private set; }
+
+        public string Description { get; private set; }
+
+        public string Unit { get; private set; }
+
+        public Meter Meter { get; private set; }
 
         public DateTimeOffset StartTimeExclusive { get; private set; }
 
@@ -40,20 +47,22 @@ namespace OpenTelemetry.Metrics
 
         public IDataValue LastValue => this.value;
 
-        public void Init(string name, DateTimeOffset startTimeExclusive, KeyValuePair<string, object>[] attributes)
+        public void Init(string name, string description, string unit, Meter meter, DateTimeOffset startTimeExclusive, KeyValuePair<string, object>[] attributes)
         {
             this.Name = name;
+            this.Description = description;
+            this.Unit = unit;
+            this.Meter = meter;
             this.StartTimeExclusive = startTimeExclusive;
             this.EndTimeInclusive = startTimeExclusive;
             this.Attributes = attributes;
         }
 
-        public void Update<T>(DateTimeOffset dt, T value)
+        public void Update<T>(T value)
             where T : struct
         {
             lock (this.lockUpdate)
             {
-                this.EndTimeInclusive = dt;
                 this.value = new DataValue<T>(value);
             }
         }
@@ -64,7 +73,7 @@ namespace OpenTelemetry.Metrics
 
             lock (this.lockUpdate)
             {
-                cloneItem.Init(this.Name, this.StartTimeExclusive, this.Attributes);
+                cloneItem.Init(this.Name, this.Description, this.Unit, this.Meter, this.StartTimeExclusive, this.Attributes);
                 cloneItem.Exemplars = this.Exemplars;
                 cloneItem.EndTimeInclusive = dt;
                 cloneItem.value = this.LastValue;