+++ src/OpenTelemetry/Metrics/MetricAggregators/HistogramMetricAggregator.cs
@@ -16,6 +16,7 @@
 
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.Metrics;
 
 namespace OpenTelemetry.Metrics
 {
@@ -25,7 +26,7 @@ namespace OpenTelemetry.Metrics
         private List<HistogramBucket> buckets = new List<HistogramBucket>();
         private double[] bounds;
 
-        public HistogramMetricAggregator(bool isDelta, double[] bounds)
+        internal HistogramMetricAggregator(bool isDelta, double[] bounds)
         {
             this.IsDeltaTemporality = isDelta;
             this.bounds = bounds;
@@ -33,13 +34,19 @@ namespace OpenTelemetry.Metrics
 
         public string Name { get; private set; }
 
+        public string Description { get; private set; }
+
+        public string Unit { get; private set; }
+
+        public Meter Meter { get; private set; }
+
         public DateTimeOffset StartTimeExclusive { get; private set; }
 
         public DateTimeOffset EndTimeInclusive { get; private set; }
 
         public KeyValuePair<string, object>[] Attributes { get; private set; }
 
-        public bool IsDeltaTemporality { get; }
+        public bool IsDeltaTemporality { get; private set; }
 
         public IEnumerable<IExemplar> Exemplars { get; private set; } = new List<IExemplar>();
 
@@ -49,22 +56,24 @@ namespace OpenTelemetry.Metrics
 
         public IEnumerable<HistogramBucket> Buckets => this.buckets;
 
-        public void Init(string name, DateTimeOffset startTimeExclusive, KeyValuePair<string, object>[] attributes)
+        public void Init(string name, string description, string unit, Meter meter, DateTimeOffset startTimeExclusive, KeyValuePair<string, object>[] attributes)
         {
             this.Name = name;
+            this.Description = description;
+            this.Unit = unit;
+            this.Meter = meter;
             this.StartTimeExclusive = startTimeExclusive;
             this.EndTimeInclusive = startTimeExclusive;
             this.Attributes = attributes;
         }
 
-        public void Update<T>(DateTimeOffset dt, T value)
+        public void Update<T>(T value)
             where T : struct
         {
             // TODO: Implement Histogram!
 
             lock (this.lockUpdate)
             {
-                this.EndTimeInclusive = dt;
                 this.PopulationCount++;
             }
         }
@@ -81,16 +90,20 @@ namespace OpenTelemetry.Metrics
 
             lock (this.lockUpdate)
             {
-                cloneItem.Init(this.Name, this.StartTimeExclusive, this.Attributes);
+                cloneItem.Init(this.Name, this.Description, this.Unit, this.Meter, this.StartTimeExclusive, this.Attributes);
                 cloneItem.Exemplars = this.Exemplars;
                 cloneItem.EndTimeInclusive = dt;
                 cloneItem.PopulationCount = this.PopulationCount;
                 cloneItem.PopulationSum = this.PopulationSum;
                 cloneItem.buckets = this.buckets;
+                cloneItem.IsDeltaTemporality = this.IsDeltaTemporality;
 
-                this.StartTimeExclusive = dt;
-                this.PopulationCount = 0;
-                this.PopulationSum = 0;
+                if (this.IsDeltaTemporality)
+                {
+                    this.StartTimeExclusive = dt;
+                    this.PopulationCount = 0;
+                    this.PopulationSum = 0;
+                }
             }
 
             return cloneItem;