+++ src/OpenTelemetry.Exporter.Console/CHANGELOG.md
@@ -9,6 +9,17 @@ please check the latest changes
 
 ## Unreleased
 
+## 1.1.0
+
+Released 2021-Jul-12
+
+* Supports OpenTelemetry.Extensions.Hosting based configuration for
+  of `ConsoleExporterOptions`.
+
+## 1.1.0-rc1
+
+Released 2021-Jun-25
+
 ## 1.1.0-beta4
 
 Released 2021-Jun-09