+++ src/OpenTelemetry.Exporter.InMemory/CHANGELOG.md
@@ -9,6 +9,17 @@ please check the latest changes
 
 ## Unreleased
 
+## 1.1.0
+
+Released 2021-Jul-12
+
+* Supports OpenTelemetry.Extensions.Hosting based configuration for `InMemoryExporter`
+  ([#2129](https://github.com/open-telemetry/opentelemetry-dotnet/pull/2129))
+
+## 1.1.0-rc1
+
+Released 2021-Jun-25
+
 ## 1.1.0-beta4
 
 Released 2021-Jun-09